requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
bs4
langchain-text-splitters
langchain-openai
chromadb
langchain_chroma
langchain-groq
duckduckgo_search==5.3.1b1
validators==0.28.1
fastapi
langserve
sse_starlette
streamlit